Rename and Clean Up Test Data
Moved around some test data to make it more readable.
Since `otherData` was renamed to `commonData` it makes
more sense to move any data that is only used in one
place to a more accurately named file.

Moved control prose data to ControlsData.js
Moved profile test data to new file ProfileData.js
